back paper/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyPDF2==3.0.1
pylatexenc==2.10
openai==1.12.0
python-pptx==0.6.23
Pillow==10.2.0
moviepy==1.0.3